refactor: merge NTT modules; rename prover.rs to fibonacci.rs
Two structural cleanups, no behaviour change:

1. Collapse src/ntt_babybear.rs and src/cuda_ntt.rs into a single
   src/ntt.rs. The CPU NTT lives at the top level of the module; the
   CUDA implementation is in an inline private submodule and its
   public surface (cuda_available, ntt_cuda, intt_cuda, CudaBuffer,
   cuda_ntt_batched, cuda_intt_batched, cuda_rs_encode_vertical) is
   re-exported flat so call sites don't have to know about the split.
   math/domain.rs is updated to use crate::ntt instead of the two
   former modules.

2. Rename src/prover.rs to src/fibonacci.rs. The contents are the
   STARK prover for the toy fibonacci AIR — it isn't the prover the
   zkvm crate uses, and the old name implied otherwise. verifier.rs
   (still fibonacci-specific) updates its import accordingly. The
   in-file test now uses super::StarkProver.

All 28 existing unit tests still pass. zkvm builds unchanged.
